// 06-miportfolio/src/pages/Ordenadores/interfaces/IOrdenador.ts
//Para poder definir el tipo de datos de un array, se pega el mismo usando PASTE AS JSON pero debe llevar las comillas en 
//cada categoría o características del array, quedaría del modo que se ve a continuación

// Generated by https://quicktype.io

export interface IOrdenador {
    nombre:          string;
    marca:           string;
    descripcion:     string;
    caracteristicas: ICaracteristicas;
    precio:          IPrecio;
    imagenes:        any[];
    valoraciones:    IValoraciones;
}

export interface ICaracteristicas {
    procesador: string;
    ram:        string;
    hdd:        string;
}

export interface IPrecio {
    pReal:    number;
    dto:      number;
    oferta:   boolean;
    cantidad: number;
}

export interface IValoraciones {
    positivas: number;
    negativas: number;
}